+++ package.json
@@ -10,6 +10,7 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.3.1",
     "@react-navigation/native": "^6.0.10",
     "@react-navigation/native-stack": "^6.6.2",
     "expo": "~45.0.0",
@@ -17,8 +18,8 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
-    "react-native-safe-area-context": "^4.3.1",
-    "react-native-screens": "^3.13.1",
+    "react-native-safe-area-context": "^4.2.4",
+    "react-native-screens": "~3.11.1",
     "react-native-web": "0.17.7"
   },
   "devDependencies": {
