+++ services/ProductsService.js
@@ -3,21 +3,21 @@ const PRODUCTS = [
         id: 100,
         name: 'ReactProduct1',
         price: 350,
-        image: require('../assets/icon.png'),
+        image: require('../assets/product.jpg'),
         description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse et rutrum augue. Nullam sapien diam, feugiat a feugiat sit amet, mollis eget nisi.'
     },
     {
         id: 101,
         name: 'Prod2',
         price: 600,
-        image: require('../assets/icon.png'),
+        image: require('../assets/product.jpg'),
         description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse et rutrum augue. Nullam sapien diam, feugiat a feugiat sit amet, mollis eget nisi.'
     },
     {
         id: 102,
         name: 'Cupcake',
         price: 2,
-        image: require('../assets/icon.png'),
+        image: require('../assets/product.jpg'),
         description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse et rutrum augue. Nullam sapien diam, feugiat a feugiat sit amet, mollis eget nisi.'
     }
 ];
